examples_v1/files_transform/pyproject.toml
[project]
name = "files-transform"
version = "0.1.0"
description = "Simple example for cocoindex: transform files in a directory."
requires-python = ">=3.11"
dependencies = ["cocoindex>=0.3.9", "markdown-it-py[linkify,plugins]"]

[tool.setuptools]
packages = []
